Refactor CSL-YAML generation
